refactor: move registry.lua to manage
